requirements.txt
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0
configparser==6.0.0